Refactor branded controls into focused files
