+++ Sources/Capsomnia/LEDToggle.swift
@@ -1,0 +1,110 @@
+import AppKit
+
+/// On/off pill toggle drawn in the landing-page LED palette.
+final class LEDToggle: NSView {
+    private let track = CALayer()
+    private let knob = CALayer()
+    private(set) var isOn: Bool
+    var onToggle: ((Bool) -> Void)?
+
+    init(isOn: Bool) {
+        self.isOn = isOn
+        super.init(frame: NSRect(x: 0, y: 0, width: 42, height: 24))
+        wantsLayer = true
+        translatesAutoresizingMaskIntoConstraints = false
+        widthAnchor.constraint(equalToConstant: 42).isActive = true
+        heightAnchor.constraint(equalToConstant: 24).isActive = true
+
+        track.frame = NSRect(x: 0, y: 0, width: 42, height: 24)
+        track.cornerRadius = 12
+        layer?.addSublayer(track)
+
+        knob.frame = NSRect(x: 3, y: 3, width: 18, height: 18)
+        knob.cornerRadius = 9
+        knob.backgroundColor = NSColor.white.cgColor
+        knob.shadowColor = NSColor.black.cgColor
+        knob.shadowOpacity = 0.35
+        knob.shadowRadius = 2
+        knob.shadowOffset = CGSize(width: 0, height: -1)
+        layer?.addSublayer(knob)
+
+        setAccessibilityElement(true)
+        setAccessibilityRole(.checkBox)
+        setAccessibilityEnabled(true)
+        focusRingType = .exterior
+        apply(animated: false)
+    }
+
+    required init?(coder: NSCoder) { nil }
+
+    override var acceptsFirstResponder: Bool {
+        true
+    }
+
+    func setOn(_ value: Bool) {
+        isOn = value
+        apply(animated: false)
+    }
+
+    override func mouseDown(with event: NSEvent) {
+        window?.makeFirstResponder(self)
+        toggle()
+    }
+
+    override func keyDown(with event: NSEvent) {
+        if event.keyCode == 36 || event.charactersIgnoringModifiers == " " {
+            toggle()
+        } else {
+            super.keyDown(with: event)
+        }
+    }
+
+    override func accessibilityPerformPress() -> Bool {
+        toggle()
+        return true
+    }
+
+    override func becomeFirstResponder() -> Bool {
+        let result = super.becomeFirstResponder()
+        needsDisplay = true
+        return result
+    }
+
+    override func resignFirstResponder() -> Bool {
+        let result = super.resignFirstResponder()
+        needsDisplay = true
+        return result
+    }
+
+    override func drawFocusRingMask() {
+        NSBezierPath(
+            roundedRect: bounds,
+            xRadius: bounds.height / 2,
+            yRadius: bounds.height / 2
+        ).fill()
+    }
+
+    override var focusRingMaskBounds: NSRect {
+        bounds
+    }
+
+    private func toggle() {
+        isOn.toggle()
+        apply(animated: true)
+        onToggle?(isOn)
+    }
+
+    override func resetCursorRects() {
+        addCursorRect(bounds, cursor: .pointingHand)
+    }
+
+    private func apply(animated: Bool) {
+        CATransaction.begin()
+        CATransaction.setDisableActions(!animated)
+        CATransaction.setAnimationDuration(0.18)
+        track.backgroundColor = (isOn ? Brand.led : Brand.offDot).cgColor
+        knob.frame.origin.x = isOn ? 21 : 3
+        CATransaction.commit()
+        setAccessibilityValue(isOn ? 1 : 0)
+    }
+}
