+++ Sources/Capsomnia/ShortcutRecorderButton.swift
@@ -1,319 +1,5 @@
 import AppKit
 
-// MARK: - Branded controls
-
-/// On/off pill toggle drawn in the landing-page LED palette.
-final class LEDToggle: NSView {
-    private let track = CALayer()
-    private let knob = CALayer()
-    private(set) var isOn: Bool
-    var onToggle: ((Bool) -> Void)?
-
-    init(isOn: Bool) {
-        self.isOn = isOn
-        super.init(frame: NSRect(x: 0, y: 0, width: 42, height: 24))
-        wantsLayer = true
-        translatesAutoresizingMaskIntoConstraints = false
-        widthAnchor.constraint(equalToConstant: 42).isActive = true
-        heightAnchor.constraint(equalToConstant: 24).isActive = true
-
-        track.frame = NSRect(x: 0, y: 0, width: 42, height: 24)
-        track.cornerRadius = 12
-        layer?.addSublayer(track)
-
-        knob.frame = NSRect(x: 3, y: 3, width: 18, height: 18)
-        knob.cornerRadius = 9
-        knob.backgroundColor = NSColor.white.cgColor
-        knob.shadowColor = NSColor.black.cgColor
-        knob.shadowOpacity = 0.35
-        knob.shadowRadius = 2
-        knob.shadowOffset = CGSize(width: 0, height: -1)
-        layer?.addSublayer(knob)
-
-        setAccessibilityElement(true)
-        setAccessibilityRole(.checkBox)
-        setAccessibilityEnabled(true)
-        focusRingType = .exterior
-        apply(animated: false)
-    }
-
-    required init?(coder: NSCoder) { nil }
-
-    override var acceptsFirstResponder: Bool {
-        true
-    }
-
-    func setOn(_ value: Bool) {
-        isOn = value
-        apply(animated: false)
-    }
-
-    override func mouseDown(with event: NSEvent) {
-        window?.makeFirstResponder(self)
-        toggle()
-    }
-
-    override func keyDown(with event: NSEvent) {
-        if event.keyCode == 36 || event.charactersIgnoringModifiers == " " {
-            toggle()
-        } else {
-            super.keyDown(with: event)
-        }
-    }
-
-    override func accessibilityPerformPress() -> Bool {
-        toggle()
-        return true
-    }
-
-    override func becomeFirstResponder() -> Bool {
-        let result = super.becomeFirstResponder()
-        needsDisplay = true
-        return result
-    }
-
-    override func resignFirstResponder() -> Bool {
-        let result = super.resignFirstResponder()
-        needsDisplay = true
-        return result
-    }
-
-    override func drawFocusRingMask() {
-        NSBezierPath(
-            roundedRect: bounds,
-            xRadius: bounds.height / 2,
-            yRadius: bounds.height / 2
-        ).fill()
-    }
-
-    override var focusRingMaskBounds: NSRect {
-        bounds
-    }
-
-    private func toggle() {
-        isOn.toggle()
-        apply(animated: true)
-        onToggle?(isOn)
-    }
-
-    override func resetCursorRects() {
-        addCursorRect(bounds, cursor: .pointingHand)
-    }
-
-    private func apply(animated: Bool) {
-        CATransaction.begin()
-        CATransaction.setDisableActions(!animated)
-        CATransaction.setAnimationDuration(0.18)
-        track.backgroundColor = (isOn ? Brand.led : Brand.offDot).cgColor
-        knob.frame.origin.x = isOn ? 21 : 3
-        CATransaction.commit()
-        setAccessibilityValue(isOn ? 1 : 0)
-    }
-}
-
-/// A compact, native pop-up button for choosing one language.
-final class LanguagePopUpButton: NSPopUpButton {
-    var onSelect: ((String) -> Void)?
-
-    init(items: [(title: String, value: String)], selected: String) {
-        super.init(frame: .zero, pullsDown: false)
-        translatesAutoresizingMaskIntoConstraints = false
-        setContentHuggingPriority(.required, for: .horizontal)
-        setContentCompressionResistancePriority(.required, for: .horizontal)
-        controlSize = .small
-        font = .systemFont(ofSize: 12, weight: .semibold)
-        alignment = .left
-        bezelStyle = .rounded
-        bezelColor = Brand.surface2
-        contentTintColor = Brand.text
-        target = self
-        action = #selector(selectionChanged)
-
-        menu?.removeAllItems()
-        for item in items {
-            let menuItem = NSMenuItem(title: item.title, action: nil, keyEquivalent: "")
-            menuItem.representedObject = item.value
-            menu?.addItem(menuItem)
-        }
-        setSelected(selected)
-
-        NSLayoutConstraint.activate([
-            widthAnchor.constraint(equalToConstant: 124),
-            heightAnchor.constraint(equalToConstant: 28)
-        ])
-    }
-
-    required init?(coder: NSCoder) { nil }
-
-    var selectedValue: String {
-        selectedItem?.representedObject as? String ?? ""
-    }
-
-    func setSelected(_ value: String) {
-        guard let item = itemArray.first(where: { $0.representedObject as? String == value }) else {
-            selectItem(at: 0)
-            return
-        }
-        select(item)
-    }
-
-    @objc private func selectionChanged() {
-        onSelect?(selectedValue)
-    }
-}
-
-/// A full-width navigation card used to reveal a deeper settings page.
-final class DisclosureButton: NSView {
-    private let titleLabel = NSTextField(labelWithString: "")
-    private let chevronHolder = NSView()
-    private let chevronView = NSImageView()
-    private var isHovered = false
-
-    var onClick: (() -> Void)?
-
-    init() {
-        super.init(frame: .zero)
-
-        translatesAutoresizingMaskIntoConstraints = false
-        wantsLayer = true
-        layer?.cornerRadius = 12
-        layer?.borderWidth = 1
-        focusRingType = .exterior
-
-        setAccessibilityElement(true)
-        setAccessibilityRole(.button)
-        setAccessibilityEnabled(true)
-
-        titleLabel.font = .systemFont(ofSize: 13, weight: .semibold)
-        titleLabel.textColor = Brand.text
-        titleLabel.lineBreakMode = .byTruncatingTail
-
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        titleLabel.setContentHuggingPriority(.defaultLow, for: .horizontal)
-        titleLabel.setContentCompressionResistancePriority(.defaultLow, for: .horizontal)
-
-        chevronHolder.translatesAutoresizingMaskIntoConstraints = false
-        chevronHolder.wantsLayer = true
-        chevronHolder.layer?.cornerRadius = 9
-
-        chevronView.image = NSImage(
-            systemSymbolName: "chevron.forward",
-            accessibilityDescription: nil
-        )
-        chevronView.symbolConfiguration = NSImage.SymbolConfiguration(
-            pointSize: 11,
-            weight: .semibold
-        )
-        chevronView.translatesAutoresizingMaskIntoConstraints = false
-        chevronHolder.addSubview(chevronView)
-
-        let content = NSStackView(views: [titleLabel, chevronHolder])
-        content.orientation = .horizontal
-        content.alignment = .centerY
-        content.distribution = .fill
-        content.spacing = 12
-        content.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(content)
-
-        NSLayoutConstraint.activate([
-            heightAnchor.constraint(equalToConstant: 52),
-            chevronHolder.widthAnchor.constraint(equalToConstant: 30),
-            chevronHolder.heightAnchor.constraint(equalToConstant: 30),
-            chevronView.centerXAnchor.constraint(equalTo: chevronHolder.centerXAnchor),
-            chevronView.centerYAnchor.constraint(equalTo: chevronHolder.centerYAnchor),
-            content.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 16),
-            content.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -12),
-            content.centerYAnchor.constraint(equalTo: centerYAnchor)
-        ])
-
-        addTrackingArea(NSTrackingArea(
-            rect: .zero,
-            options: [.activeInActiveApp, .inVisibleRect, .mouseEnteredAndExited],
-            owner: self
-        ))
-        refreshAppearance()
-    }
-
-    required init?(coder: NSCoder) { nil }
-
-    override var acceptsFirstResponder: Bool {
-        true
-    }
-
-    func setTitle(_ title: String) {
-        titleLabel.stringValue = title
-        setAccessibilityLabel(title)
-        setAccessibilityHelp(nil)
-    }
-
-    override func mouseDown(with event: NSEvent) {
-        window?.makeFirstResponder(self)
-        onClick?()
-    }
-
-    override func mouseEntered(with event: NSEvent) {
-        isHovered = true
-        refreshAppearance()
-    }
-
-    override func mouseExited(with event: NSEvent) {
-        isHovered = false
-        refreshAppearance()
-    }
-
-    override func keyDown(with event: NSEvent) {
-        if event.keyCode == 36 || event.charactersIgnoringModifiers == " " {
-            onClick?()
-        } else {
-            super.keyDown(with: event)
-        }
-    }
-
-    override func accessibilityPerformPress() -> Bool {
-        onClick?()
-        return true
-    }
-
-    override func becomeFirstResponder() -> Bool {
-        let result = super.becomeFirstResponder()
-        needsDisplay = true
-        return result
-    }
-
-    override func resignFirstResponder() -> Bool {
-        let result = super.resignFirstResponder()
-        needsDisplay = true
-        return result
-    }
-
-    override func drawFocusRingMask() {
-        NSBezierPath(
-            roundedRect: bounds,
-            xRadius: 12,
-            yRadius: 12
-        ).fill()
-    }
-
-    override var focusRingMaskBounds: NSRect {
-        bounds
-    }
-
-    override func resetCursorRects() {
-        addCursorRect(bounds, cursor: .pointingHand)
-    }
-
-    private func refreshAppearance() {
-        layer?.backgroundColor = (isHovered ? Brand.surface2 : Brand.surface).cgColor
-        layer?.borderColor = (
-            isHovered ? Brand.led.withAlphaComponent(0.38) : Brand.border
-        ).cgColor
-        chevronHolder.layer?.backgroundColor = (
-            isHovered ? Brand.led.withAlphaComponent(0.13) : Brand.surface2
-        ).cgColor
-        chevronView.contentTintColor = isHovered ? Brand.led : Brand.textDim
-    }
-}
-
 /// A keycap-style recorder for Capsomnia's persisted global shortcut.
 final class ShortcutRecorderButton: NSView {
     private var placeholderTitle: String
@@ -703,112 +389,4 @@ private final class ShortcutKeycapView: NSView {
     }
 
     required init?(coder: NSCoder) { nil }
-}
-
-/// LED-green primary button matching the landing-page CTA.
-final class LEDButton: NSView {
-    private let label = NSTextField(labelWithString: "")
-    var onClick: (() -> Void)?
-
-    var title: String {
-        get { label.stringValue }
-        set {
-            label.stringValue = newValue
-            setAccessibilityLabel(newValue)
-        }
-    }
-
-    init() {
-        super.init(frame: .zero)
-        wantsLayer = true
-        layer?.cornerRadius = 11
-        layer?.backgroundColor = Brand.led.cgColor
-        translatesAutoresizingMaskIntoConstraints = false
-        setAccessibilityElement(true)
-        setAccessibilityRole(.button)
-        setAccessibilityEnabled(true)
-        focusRingType = .exterior
-
-        label.font = .systemFont(ofSize: 13, weight: .semibold)
-        label.textColor = .black
-        label.alignment = .center
-        label.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(label)
-
-        NSLayoutConstraint.activate([
-            heightAnchor.constraint(equalToConstant: 38),
-            label.centerXAnchor.constraint(equalTo: centerXAnchor),
-            label.centerYAnchor.constraint(equalTo: centerYAnchor)
-        ])
-
-        addTrackingArea(NSTrackingArea(
-            rect: .zero,
-            options: [.activeInActiveApp, .inVisibleRect, .mouseEnteredAndExited],
-            owner: self
-        ))
-    }
-
-    required init?(coder: NSCoder) { nil }
-
-    override var acceptsFirstResponder: Bool {
-        true
-    }
-
-    override func mouseDown(with event: NSEvent) {
-        window?.makeFirstResponder(self)
-        performAction()
-    }
-
-    override func keyDown(with event: NSEvent) {
-        if event.keyCode == 36 || event.charactersIgnoringModifiers == " " {
-            performAction()
-        } else {
-            super.keyDown(with: event)
-        }
-    }
-
-    override func accessibilityPerformPress() -> Bool {
-        performAction()
-        return true
-    }
-
-    override func becomeFirstResponder() -> Bool {
-        let result = super.becomeFirstResponder()
-        needsDisplay = true
-        return result
-    }
-
-    override func resignFirstResponder() -> Bool {
-        let result = super.resignFirstResponder()
-        needsDisplay = true
-        return result
-    }
-
-    override func drawFocusRingMask() {
-        NSBezierPath(
-            roundedRect: bounds,
-            xRadius: 11,
-            yRadius: 11
-        ).fill()
-    }
-
-    override var focusRingMaskBounds: NSRect {
-        bounds
-    }
-
-    override func mouseEntered(with event: NSEvent) {
-        layer?.backgroundColor = Brand.ledBright.cgColor
-    }
-
-    override func mouseExited(with event: NSEvent) {
-        layer?.backgroundColor = Brand.led.cgColor
-    }
-
-    override func resetCursorRects() {
-        addCursorRect(bounds, cursor: .pointingHand)
-    }
-
-    private func performAction() {
-        onClick?()
-    }
 }
